refactor ReflectionUtil to util.reflection package
