+++ common/api/src/main/java/com/dfsek/terra/api/event/events/config/ConfigurationLoadEvent.java
@@ -5,7 +5,7 @@ import com.dfsek.tectonic.config.ConfigTemplate;
 import com.dfsek.terra.api.config.ConfigPack;
 import com.dfsek.terra.api.config.ConfigType;
 import com.dfsek.terra.api.event.events.PackEvent;
-import com.dfsek.terra.api.util.ReflectionUtil;
+import com.dfsek.terra.api.util.reflection.ReflectionUtil;
 
 import java.util.function.Consumer;
