+++ common/api/src/main/java/com/dfsek/terra/api/util/reflection/ReflectionUtil.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.api.util;
+package com.dfsek.terra.api.util.reflection;
 
 import org.jetbrains.annotations.NotNull;
