+++ common/api/src/main/java/com/dfsek/terra/api/util/reflection/TypeKey.java
@@ -1,7 +1,5 @@
 package com.dfsek.terra.api.util.reflection;
 
-import com.dfsek.terra.api.util.ReflectionUtil;
-
 import java.lang.reflect.*;
 import java.util.Arrays;
 import java.util.Objects;
