+++ common/implementation/src/main/java/com/dfsek/terra/commands/TerraCommandManager.java
@@ -22,7 +22,7 @@ import com.dfsek.terra.api.command.tab.TabCompleter;
 import com.dfsek.terra.api.entity.CommandSender;
 import com.dfsek.terra.api.entity.Player;
 import com.dfsek.terra.api.injection.exception.InjectionException;
-import com.dfsek.terra.api.util.ReflectionUtil;
+import com.dfsek.terra.api.util.reflection.ReflectionUtil;
 import com.dfsek.terra.inject.InjectorImpl;
 import net.jafama.FastMath;
